IP Support update for 2.3.0 release
This release contains:
- Support for Mali-G627, Mali-G725 and Immortalis-G925 GPUs
- Improved counter naming and descriptions for queue events.

+++ backend/device/src/device/hwcnt/sampler/kinstr_prfcnt/backend.hpp
@@ -155,7 +155,7 @@ class backend : public base::backend<syscall_iface_t> {
     }
 
     /** Size of metadata item: includes sample metadata, clock metadata and block metadata. */
-    const uint32_t metadata_item_size_{};
+    const size_t metadata_item_size_{};
     /** Block index remap instance, if any. */
     const block_index_remap *remap_;
 };
